+++ src/components/Navbar/Search/Category.module.scss
@@ -1,0 +1,104 @@
+@import '~@santiment-network/ui/mixins';
+
+.category {
+  display: flex;
+  flex-direction: column;
+  margin-right: 8px;
+  flex: 1;
+
+  &:last-child {
+    margin: 0;
+  }
+}
+
+.loader {
+  $border: 2.5px;
+
+  display: inline-block;
+  width: 12px;
+  height: 12px;
+  border-radius: 50%;
+  margin-right: 6px;
+  border-top: $border solid var(--mystic);
+  border-right: $border solid var(--mystic);
+  border-bottom: $border solid var(--mystic);
+  border-left: $border solid var(--casper);
+  animation: loading 1.1s infinite linear;
+  margin-left: 6px;
+}
+
+@keyframes loading {
+  0% {
+    transform: rotate(0deg);
+  }
+
+  100% {
+    transform: rotate(360deg);
+  }
+}
+
+.title {
+  @include text('body-3', 'm');
+
+  color: var(--waterloo);
+  border-bottom: 1px solid var(--porcelain);
+  margin: 0 8px 10px;
+}
+
+.button {
+  @include text('body-3');
+
+  outline: none;
+  border: none;
+  min-height: 32px;
+  width: 100%;
+  color: var(--rhino);
+  background: none;
+  cursor: pointer;
+  border-radius: 4px;
+  text-align: left;
+  padding: 6px 8px;
+  display: flex;
+  align-items: center;
+
+  &:hover {
+    background: var(--athens);
+
+    &,
+    & * {
+      color: var(--jungle-green);
+    }
+  }
+}
+
+// Assets
+.ticker {
+  color: var(--waterloo);
+  margin-left: 4px;
+}
+
+// People
+.avatar,
+.fallback {
+  width: 20px;
+  height: 20px;
+  border-radius: 50%;
+  margin-right: 10px;
+}
+
+.fallback {
+  background: var(--porcelain);
+  color: var(--waterloo) !important;
+  font-weight: 600;
+  font-size: 10px;
+  text-align: center;
+}
+
+// Insights
+.category_insights {
+  min-width: 280px;
+}
+
+.icon {
+  margin-right: 6px;
+}
