+++ src/components/Navbar/Search/Suggestions.module.scss
@@ -1,104 +1,38 @@
-@import '~@santiment-network/ui/mixins';
+$duration: 250;
+$delay: 80;
 
-.title {
-  @include text('body-3', 'm');
-
-  color: var(--waterloo);
-  border-bottom: 1px solid var(--porcelain);
-  margin: 0 8px 10px;
-}
-
-.category {
+.dropdown {
+  background: var(--white);
+  border: 1px solid var(--porcelain);
+  box-shadow: 0 4px 8px rgba(24, 27, 43, 0.03);
+  border-radius: 4px;
+  padding: 16px;
   display: flex;
-  flex-direction: column;
-  margin-right: 8px;
-  flex: 1;
-
-  &:last-child {
-    margin: 0;
-  }
+  min-width: 100%;
+  z-index: 10;
+  position: absolute;
+  right: 0;
+  top: calc(100% + 8px);
+  transform-origin: top center;
 }
 
-.loader {
-  $border: 2.5px;
-
-  display: inline-block;
-  width: 12px;
-  height: 12px;
-  border-radius: 50%;
-  margin-right: 6px;
-  border-top: $border solid var(--mystic);
-  border-right: $border solid var(--mystic);
-  border-bottom: $border solid var(--mystic);
-  border-left: $border solid var(--casper);
-  animation: loading 1.1s infinite linear;
-  margin-left: 6px;
+.enter {
+  opacity: 0;
+  animation: open #{$duration}ms #{$delay}ms forwards;
 }
 
-@keyframes loading {
+.exit {
+  animation: open #{$duration}ms reverse forwards;
+}
+
+@keyframes open {
   0% {
-    transform: rotate(0deg);
+    transform: rotateX(40deg);
+    opacity: 0;
   }
 
   100% {
-    transform: rotate(360deg);
+    transform: rotateX(0deg);
+    opacity: 1;
   }
-}
-
-.button {
-  @include text('body-3');
-
-  outline: none;
-  border: none;
-  min-height: 32px;
-  width: 100%;
-  color: var(--rhino);
-  background: none;
-  cursor: pointer;
-  border-radius: 4px;
-  text-align: left;
-  padding: 6px 8px;
-  display: flex;
-  align-items: center;
-
-  &:hover {
-    background: var(--athens);
-
-    &,
-    & * {
-      color: var(--jungle-green);
-    }
-  }
-}
-
-// assets
-.ticker {
-  color: var(--waterloo);
-  margin-left: 4px;
-}
-
-// People
-.avatar,
-.fallback {
-  width: 20px;
-  height: 20px;
-  border-radius: 50%;
-  margin-right: 10px;
-}
-
-.fallback {
-  background: var(--porcelain);
-  color: var(--waterloo) !important;
-  font-weight: 600;
-  font-size: 10px;
-  text-align: center;
-}
-
-// INSIGHTS
-.category_insights {
-  min-width: 280px;
-}
-
-.icon {
-  margin-right: 6px;
 }
